Add back the tooltip for SUSE Manager's unknown state
