+++ assets/js/pages/HostDetailsPage/HostDetailsPage.jsx
@@ -30,54 +30,15 @@ import {
 
 import { deregisterHost } from '@state/hosts';
 import { fetchSoftwareUpdates } from '@state/softwareUpdates';
+
+import {
+  getSoftwareUpdatesErrorMessage,
+  getSoftwareUpdatesErrorTooltip,
+} from './suseManagerMessaging';
 import HostDetails from './HostDetails';
 
 const chartsEnabled = getFromConfig('chartsEnabled');
 const suseManagerEnabled = getFromConfig('suseManagerEnabled');
-
-const getSoftwareUpdatesErrorMessage = (errors) => {
-  const hostNotFoundInSUMA = errors.some(
-    ({ detail }) =>
-      detail === 'The requested resource cannot be found.' ||
-      detail === 'No system ID was found on SUSE Manager for this host.'
-  );
-
-  const connectionNotWorking = errors.some(
-    ({ detail }) => detail === 'Something went wrong.'
-  );
-
-  if (hostNotFoundInSUMA) {
-    return 'Host not found in SUSE Manager';
-  }
-
-  if (connectionNotWorking) {
-    return 'Connection to SUMA not working';
-  }
-
-  return 'Unknown';
-};
-
-const getSoftwareUpdatesErrorTooltip = (errors) => {
-  const hostNotFoundInSUMA = errors.some(
-    ({ detail }) =>
-      detail === 'The requested resource cannot be found.' ||
-      detail === 'No system ID was found on SUSE Manager for this host.'
-  );
-
-  const connectionNotWorking = errors.some(
-    ({ detail }) => detail === 'Something went wrong.'
-  );
-
-  if (hostNotFoundInSUMA) {
-    return 'Contact your SUSE Manager admin to ensure the host is managed by SUSE Manager';
-  }
-
-  if (connectionNotWorking) {
-    return 'Please review SUSE Manager settings';
-  }
-
-  return undefined;
-};
 
 function HostDetailsPage() {
   const { hostID } = useParams();
